first admin and role creating script added
